requirements.txt
descartes
geopandas
matplotlib
pandas
pyyaml
pyreadstat
pyarrow
loguru
scikit-learn
statsmodels